sentiment-analysis/requirements.txt
aniso8601==6.0.0
astroid==2.2.5
beautifulsoup4==4.7.1
boto3==1.9.113
botocore==1.12.113
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
cycler==0.10.0
docutils==0.14
Flask==0.12.1
Flask-RESTful==0.3.5
google==2.0.2
google-api-core==1.8.1
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-cloud-language==1.1.1
google-cloud-speech==0.36.3
googleapis-common-protos==1.5.8
grpcio==1.19.0
httplib2==0.12.1
idna==2.8
isort==4.3.15
itsdangerous==1.1.0
Jinja2==2.10
jmespath==0.9.4
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mccabe==0.6.1
nltk==3.4
numpy==1.16.2
oauthlib==3.0.1
pandas==0.24.1
protobuf==3.7.0
py==1.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.4
pylint==2.3.1
pyparsing==2.3.1
PySocks==1.6.8
pytest==3.0.7
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
rsa==4.0
s3transfer==0.2.0
singledispatch==3.4.0.3
six==1.12.0
soupsieve==1.8
textblob==0.15.3
tweepy==3.7.0
typed-ast==1.3.1
uritemplate==3.0.0
urllib3==1.24.1
Werkzeug==0.14.1
wrapt==1.11.1

